REFACTOR layout changes + styling
